Add files via upload
We present here three applications in simulation. All of them were proposed by Simulation cours at ISIMA france in 2014. 

In (S_BankAtt) we simulate cashier attendence. This is a simplified version of the reality because we have just one cashier. More over the cashier has a queue limit. It means that it can not receive more clients when the queue if full in potencial clients that arrive will be lost. (To run you just need to compile and run)


In (S_MachineParalle) we simulate a system with 3 machine different (2 in parallel). The objective of this simulation is to calculate the number of pieces produced for the specified conditions. (To run this application in your computer you need a visual Studio Version greater than 8). 


The (SimulationRobot) simulates a automized production line. Where different pieces must pass in different machines and different order. Moreover, we ilustrate this simulation in our application  (To run this application in your computer you need a visual Studio Version greater than 8).