Update WordCounter.cpp
(WordCounter.cpp) -Objective: This application aims to rank the most used words in a file. -Headers: No headers used. -Classes: We are using the Standard Containers "map" and "set". The "map" help us to store/find words, and "set" is used at the end to rank our words. Moreover, we created a class named as aWORD that store the information of each word (string and counter), this class is used to rank our words. -Compilation To compile on Linux type: make WordCounter_make -Run There are two ways of running this application: a - By command line arguments: "./WordCounter myFile.something topList" where "myFile" is the name of the file that you want to read, and "topList" is an integer number to the size of the top most used words to be printed on the screen. b - Without command line: "./WordCounter" This way, the application will ask you the "file" and "topList" during the execution.